Move audio traits into Sounds namespace.
