+++ OpenRA.Mods.Common/Traits/Sound/AmbientSound.cs
@@ -11,7 +11,7 @@
 
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.Common.Traits
+namespace OpenRA.Mods.Common.Traits.Sound
 {
 	[Desc("Plays a looping audio file at the actor position. Attach this to the `World` actor to cover the whole map.")]
 	class AmbientSoundInfo : UpgradableTraitInfo
